Rename component input-project-info
